+++ Source/Jeeves/Settings.cs
@@ -1,18 +1,14 @@
-﻿using Jeeves.Core;
-
-namespace Jeeves
+﻿namespace Jeeves
 {
     internal sealed class Settings
     {
-        public Settings(string databasePath, string baseUrl, SecurityOption security)
+        public Settings(string databasePath, string baseUrl)
         {
             DatabasePath = databasePath;
             BaseUrl = baseUrl;
-            Security = security;
         }
 
         public string DatabasePath { get; }
         public string BaseUrl { get; }
-        public SecurityOption Security { get; }
     }
 }